data_fetcher/requirements.txt
requests>=2.28
pymongo>=4.0
python-dotenv>=1.0
beautifulsoup4>=4.12
python-dateutil>=2.8
